day09/lab/requirements.txt
# Day 09 Lab — Multi-Agent Orchestration
# Install: pip install -r requirements.txt

# ─── Core ───────────────────────────────
python-dotenv>=1.0.0

# ─── Vector Store ───────────────────────
chromadb>=0.4.22

# ─── Embeddings ─────────────────────────
# Option A: Sentence Transformers (offline, không cần API key)
sentence-transformers>=2.6.0

# ─── LLM ────────────────────────────────
# Option A: OpenAI
openai>=1.12.0

# Option B: Google Gemini
# google-generativeai>=0.5.0

# ─── Orchestration ──────────────────────
# Option: LangGraph (uncomment nếu dùng LangGraph)
# langgraph>=0.1.0
# langchain>=0.2.0
# langchain-openai>=0.1.0

# ─── Data ───────────────────────────────
pyyaml>=6.0.1

# ─── MCP (Sprint 3 Bonus — HTTP Server) ───
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0
# mcp>=1.0.0      # optional: real MCP library
# httpx>=0.27.0   # optional: thay thế requests

# ─── Dev/Test ───────────────────────────
pytest>=8.0.0
